trying to fix name clash
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@5917 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/traffic_lights/MSOffTrafficLightLogic.cpp
@@ -80,8 +80,8 @@ MSOffTrafficLightLogic::getPosition(SUMOTime simStep)
     return 0;
 }
 
-size_t
-MSOffTrafficLightLogic::getStepFromPos(size_t position)
+unsigned int
+MSOffTrafficLightLogic::getStepFromPos(unsigned int position)
 {
     return 0;
 }
